WCSimSteppingAction is now an unimplemented class.
The class was not actually doing things, it just had some legacy liquid
argon code in it, that was commented it. I have removed all of this, and
as such it now just has the UserSteppingAction function, which is now
empty. Left since it might be useful in the future.