Update to add the pmt timing constant into the pmtDefinition.xml file.
The timing constant is related to the time resolution in the class
WCSimWCDigitizer in the following way:
timeRes = 0.33 + (timeConstant / hitCharge)
This equation is tuned for SK and might need to be modified.

The main change made here was to add the PMT name to the WCSimWCHit
class. This involed adding code in WCSimWCSD to strip the PMT name
from the geometry path.

Time constants have been added to the pmtDefinition.xml file, and
the current values are identical to those that existed hard-coded
in the digitizer. These will need to be update through research to
some reasonable values.